pyproject.toml
[project]
name = "codetester"
version = "0.1.0"
description = "An AI-powered code testing agent using GitHub Copilot SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "github-copilot-sdk>=0.2.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
examples = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]